iptables: add support for NOTRACK rules for pod to pod traffic
When running in KPR mode, all pod to pod traffic goes through connection
tracking twice: the first time traffic is tracked by netfilter and the
second one by the BPF datapath.

This is suboptimal as doing conntrack at the netfilter layer is not
strictly required for the operation of the CNI, but it has some
important performance implications.

This commit addresses this by installing a couple of `-j NOTRACK`
Iptables rules which will prevent netfilter from tracking pod to pod
traffic (by matching all traffic from and to the native-routing-cidr
CIDR) when Cilium is running in direct routing and strict KPR mode.

This feature can only be enabled with this specific setup as:

* when running in kube-proxy mode conntrack is required to perform NAT
* when running in tunneling mode (and with strict KPR), the pod to pod
  traffic is encapsulated and decapsulated in the context of the BPF
  datapath program and so it is already skipping the netfilter raw table
  where connection tracking happens.

Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>